# astroneko-backend/docs/swagger.yaml
basePath: /
definitions:
...
  referral_code.CreateReferralCodeRequest:
    properties:
      referral_code:
        type: string
    required:
    - referral_code
    type: object
...
  referral_code.ListReferralCodesResponse:
    properties:
      codes:
        items:
          $ref: '#/definitions/referral_code.ReferralCodeResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  referral_code.ReferralCodeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      referral_code:
        type: string
      updated_at:
        type: string
      used_count:
        type: integer
    type: object
  referral_code.UpdateReferralCodeRequest:
    properties:
      referral_code:
        type: string
    required:
    - referral_code
    type: object
...
  shared.ResponseBody:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
  shared.Status:
    properties:
      code:
        type: string
      errorID:
        type: string
      message:
        items:
          type: string
        type: array
      successID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Astroneko Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Astroneko Backend API
  version: "1.0"
paths:
...
  /v1/api/referral-codes:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all general referral codes
      parameters:
      - description: 'Number of items to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/referral_code.ListReferralCodesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: List all referral codes
      tags:
      - referral-codes
    post:
      consumes:
      - application/json
      description: Create a new general referral code (admin only)
      parameters:
      - description: Referral code data
        in: body
        name: referral_code
        required: true
        schema:
          $ref: '#/definitions/referral_code.CreateReferralCodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/referral_code.ReferralCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Create a new general referral code
      tags:
      - referral-codes
  /v1/api/referral-codes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a general referral code
      parameters:
      - description: Referral Code ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Delete referral code
      tags:
      - referral-codes
    get:
      consumes:
      - application/json
      description: Get a general referral code by its ID
      parameters:
      - description: Referral Code ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/referral_code.ReferralCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Get referral code by ID
      tags:
      - referral-codes
    put:
      consumes:
      - application/json
      description: Update a general referral code
      parameters:
      - description: Referral Code ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated referral code data
        in: body
        name: referral_code
        required: true
        schema:
          $ref: '#/definitions/referral_code.UpdateReferralCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/referral_code.ReferralCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Update referral code
      tags:
      - referral-codes
  /v1/api/referral-codes/code/{code}:
    get:
      consumes:
      - application/json
      description: Get a general referral code by its code value
      parameters:
      - description: Referral Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/referral_code.ReferralCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Get referral code by code
      tags:
      - referral-codes
  /v1/api/referral-codes/validate/{code}:
    get:
      consumes:
      - application/json
      description: Check if a referral code is valid
      parameters:
      - description: Referral Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Validate referral code
      tags:
      - referral-codes
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"